requirements.txt
Flask==3.0.2
requests==2.31.0
transformers==4.46.0
newspaper3k==0.2.8
torch==2.2.1
beautifulsoup4==4.12.3
newsapi-python==0.2.7
python-dotenv==1.0.1
gunicorn==20.1.0
numpy==1.26.4
lxml_html_clean==0.4.1